refactor: improve codebase structure and reduce duplication
- Extract isStackingContext() from context.ts into dedicated stacking-context.ts
  to reduce cognitive load (48 lines used by only static-z-index rule)
- Split usePageScan into focused modules: groupByRule() to utils/group-by-rule.ts,
  inspect logic to useInspectElement hook
- Create shared test helper make-element.ts to eliminate duplicated makeElement()
  and DEFAULT_COMPUTED_STYLES across 11 test files
- Add no-barrel-exports rule to Code Style in CLAUDE.md
- Fix documentation: add missing inline-no-vertical-margin rule, stacking-context.ts,
  and correct legacy D-1 comment in engine.test.ts
