+++ src/rules/__tests__/alignment.test.ts
@@ -1,30 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { checkAlignment } from '../alignment.ts';
 import { createRuleContext } from '../context.ts';
-import type { ElementData } from '../types.ts';
-
-function makeElement(overrides: Partial<ElementData['computedStyles']>): ElementData {
-  return {
-    tagName: 'div',
-    id: '',
-    classList: [],
-    computedStyles: {
-      display: 'block',
-      width: 'auto',
-      height: 'auto',
-      gap: 'normal',
-      rowGap: 'normal',
-      columnGap: 'normal',
-      alignItems: 'normal',
-      justifyContent: 'normal',
-      placeItems: 'normal',
-      placeContent: 'normal',
-      columnCount: 'auto',
-      ...overrides,
-    },
-    parent: null,
-  };
-}
+import { makeElement } from './helpers/make-element.ts';
 
 describe('container-no-align: alignment', () => {
   it('warns when align-items is set on block element', () => {
