+++ src/rules/__tests__/static-z-index.test.ts
@@ -2,33 +2,13 @@ import { describe, it, expect } from 'vitest';
 import { checkStaticZIndex } from '../static-z-index.ts';
 import { createRuleContext } from '../context.ts';
 import type { ElementData } from '../types.ts';
+import { makeElement as _makeElement } from './helpers/make-element.ts';
 
 function makeElement(
   styles: Partial<ElementData['computedStyles']>,
   parent: ElementData['parent'] = { computedStyles: { display: 'block' } },
 ): ElementData {
-  return {
-    tagName: 'div',
-    id: '',
-    classList: [],
-    computedStyles: {
-      display: 'block',
-      position: 'static',
-      zIndex: 'auto',
-      opacity: '1',
-      transform: 'none',
-      filter: 'none',
-      backdropFilter: 'none',
-      perspective: 'none',
-      clipPath: 'none',
-      isolation: 'auto',
-      mixBlendMode: 'normal',
-      contain: 'none',
-      willChange: 'auto',
-      ...styles,
-    },
-    parent,
-  };
+  return _makeElement(styles, parent);
 }
 
 describe('static-no-z-index', () => {
