+++ src/sidebar/hooks/__tests__/usePageScan.hook.test.tsx
@@ -187,107 +187,6 @@ describe('usePageScan (hook integration)', () => {
     expect(result.current.error).toBe('Invalid scan result');
   });
 
-  it('inspectElement success keeps inspectError null', async () => {
-    const { result } = renderHook(() => usePageScan());
-
-    act(() => {
-      result.current.inspectElement(0);
-    });
-
-    await act(async () => {
-      await mock.resolveLastEval(true);
-    });
-
-    expect(result.current.inspectError).toBeNull();
-  });
-
-  it('inspectElement failure sets error that auto-clears after 3s', async () => {
-    vi.useFakeTimers();
-    const { result } = renderHook(() => usePageScan());
-
-    act(() => {
-      result.current.inspectElement(0);
-    });
-
-    await act(async () => {
-      await mock.resolveLastEval(false);
-    });
-
-    expect(result.current.inspectError).toBe(
-      'Element not found — the page may have changed since the scan.',
-    );
-
-    // Advance past auto-clear timer
-    await act(async () => {
-      vi.advanceTimersByTime(3000);
-    });
-
-    expect(result.current.inspectError).toBeNull();
-  });
-
-  it('inspectElement clears previous timer on re-call', async () => {
-    vi.useFakeTimers();
-    const { result } = renderHook(() => usePageScan());
-
-    // First inspect fails
-    act(() => {
-      result.current.inspectElement(0);
-    });
-    await act(async () => {
-      await mock.resolveLastEval(false);
-    });
-    expect(result.current.inspectError).not.toBeNull();
-
-    // Second inspect also fails — should reset timer
-    act(() => {
-      result.current.inspectElement(1);
-    });
-    await act(async () => {
-      await mock.resolveLastEval(false);
-    });
-
-    // Advance 2s — first timer's 3s would have fired, but it was cancelled
-    await act(async () => {
-      vi.advanceTimersByTime(2000);
-    });
-    // Error should still be present (second timer hasn't fired yet)
-    expect(result.current.inspectError).not.toBeNull();
-
-    // Advance remaining 1s for second timer
-    await act(async () => {
-      vi.advanceTimersByTime(1000);
-    });
-    expect(result.current.inspectError).toBeNull();
-  });
-
-  it('inspectElement timer is cleaned up on unmount', async () => {
-    vi.useFakeTimers();
-    const { result, unmount } = renderHook(() => usePageScan());
-
-    // Trigger inspect failure — starts 3s auto-clear timer
-    act(() => {
-      result.current.inspectElement(0);
-    });
-    await act(async () => {
-      await mock.resolveLastEval(false);
-    });
-    expect(result.current.inspectError).not.toBeNull();
-
-    // Capture state before unmount
-    const errorBeforeUnmount = result.current.inspectError;
-
-    // Unmount before the 3s timer fires
-    unmount();
-
-    // Advance past the auto-clear timer — should not cause stale setState
-    await act(async () => {
-      vi.advanceTimersByTime(3000);
-    });
-
-    // State must not have changed (timer was cleared on unmount)
-    expect(result.current.inspectError).toBe(errorBeforeUnmount);
-  });
-
   it('clear() resets all state to idle', async () => {
     const { result } = renderHook(() => usePageScan());
